+++ dockerfile
@@ -1,0 +1,11 @@
+FROM python:3.9
+
+WORKDIR /app
+
+COPY requirements.txt .
+RUN pip install --no-cache-dir -r requirements.txt
+
+COPY ./app /app
+
+# Change the CMD to point to app.main
+CMD ["uvicorn", "app.main:app", "--host", "0.0.0.0", "--port", "8000"]
