+++ frontend/dockerfile
@@ -1,0 +1,16 @@
+# frontend/Dockerfile
+FROM node:16 as base
+
+WORKDIR /app
+COPY package*.json ./
+RUN npm install
+COPY . .
+
+FROM base as development
+ENV NODE_ENV=development
+CMD ["npm", "start"]
+
+FROM base as production
+ENV NODE_ENV=production
+RUN npm run build
+CMD ["npm", "start"]
