--- notifications.py
@@ -1,25 +1,0 @@
-from sqlalchemy.orm import Session
-from models import Notification, Advisory, Device
-from datetime import datetime
-
-def check_for_new_advisories(db: Session):
-    # This would be called periodically or after advisory updates
-    devices = db.query(Device).all()
-    advisories = db.query(Advisory).all()
-    
-    for device in devices:
-        for advisory in advisories:
-            if device.model.lower() in advisory.product.lower():
-                # Check if notification already exists
-                exists = db.query(Notification).filter(
-                    Notification.device_id == device.id,
-                    Notification.advisory_id == advisory.id
-                ).first()
-                
-                if not exists:
-                    notification = Notification(
-                        device_id=device.id,
-                        advisory_id=advisory.id
-                    )
-                    db.add(notification)
-    db.commit()
